:sparkles: (json) admit hex encoding
the pwm values are stored in hex format, therefore all transmited data
is passed as hex values. (base 16 for the win!)

+++ aeropendulum/esp/lib/json/inc/json_parser.h
@@ -24,7 +24,7 @@ typedef enum TokenCompareRvType {
 
 typedef struct SimpleJSONType {
     const char name[TOKEN_NAME_MAX_LEN];
-    float value;
+    uint16_t value;
 } SimpleJSONType;
